+++ torchbearer/callbacks/decorators.py
@@ -1,8 +1,7 @@
-from torchbearer.callbacks import Callback
-import torchbearer
-from torchbearer.state import StateKey
-
 from inspect import signature
+
+import torchbearer
+from torchbearer.callbacks import Callback
 
 
 class LambdaCallback(Callback):
@@ -28,22 +27,24 @@ def on_start(func):
     """ The :func:`on_start` decorator is used to initialise a :class:`.Callback` with :meth:`~.Callback.on_start`
     calling the decorated function
 
-    :param func: The function(state) to *decorate*
-    :type func: function
-    :return: Initialised callback with :meth:`~.Callback.on_start` calling func
-    :rtype: :class:`.Callback`
+    Args:
+        func (function): The function(state) to *decorate*
+
+    Returns:
+        :class:`.Callback`: Initialised callback with :meth:`~.Callback.on_start` calling func
     """
     return bind_to(Callback.on_start)(func)
 
 
 def on_start_epoch(func):
-    """ The :func:`on_start_epoch` decorator is used to initialise a :class:`.Callback` with :meth:`~.Callback.on_start_epoch`
-    calling the decorated function
+    """ The :func:`on_start_epoch` decorator is used to initialise a :class:`.Callback` with
+    :meth:`~.Callback.on_start_epoch` calling the decorated function
 
-    :param func: The function(state) to *decorate*
-    :type func: function
-    :return: Initialised callback with :meth:`~.Callback.on_start_epoch` calling func
-    :rtype: :class:`.Callback`
+        Args:
+        func (function): The function(state) to *decorate*
+
+    Returns:
+        :class:`.Callback`: Initialised callback with :meth:`~.Callback.on_start_epoch` calling func
     """
     return bind_to(Callback.on_start_epoch)(func)
 
@@ -52,10 +53,11 @@ def on_start_training(func):
     """ The :func:`on_start_training` decorator is used to initialise a :class:`.Callback` with :meth:`~.Callback.on_start_training`
     calling the decorated function
 
-    :param func: The function(state) to *decorate*
-    :type func: function
-    :return: Initialised callback with :meth:`.Callback.on_start_training` calling func
-    :rtype: :class:`.Callback`
+    Args:
+        func (function): The function(state) to *decorate*
+
+    Returns:
+        :class:`.Callback`: Initialised callback with :meth:`~.Callback.on_start_training` calling func
     """
     return bind_to(Callback.on_start_training)(func)
 
@@ -64,10 +66,11 @@ def on_sample(func):
     """ The :func:`on_sample` decorator is used to initialise a :class:`.Callback` with :meth:`~.Callback.on_sample`
     calling the decorated function
 
-    :param func: The function(state) to *decorate*
-    :type func: function
-    :return: Initialised callback with :meth:`.Callback.on_sample` calling func
-    :rtype: :class:`.Callback`
+    Args:
+        func (function): The function(state) to *decorate*
+
+    Returns:
+        :class:`.Callback`: Initialised callback with :meth:`~.Callback.on_sample` calling func
     """
     return bind_to(Callback.on_sample)(func)
 
@@ -76,10 +79,11 @@ def on_forward(func):
     """ The :func:`on_forward` decorator is used to initialise a :class:`.Callback` with :meth:`~.Callback.on_forward`
     calling the decorated function
 
-    :param func: The function(state) to *decorate*
-    :type func: function
-    :return: Initialised callback with :meth:`.Callback.on_forward` calling func
-    :rtype: :class:`.Callback`
+    Args:
+        func (function): The function(state) to *decorate*
+
+    Returns:
+        :class:`.Callback`: Initialised callback with :meth:`~.Callback.on_forward` calling func
     """
     return bind_to(Callback.on_forward)(func)
 
@@ -88,10 +92,11 @@ def on_criterion(func):
     """ The :func:`on_criterion` decorator is used to initialise a :class:`.Callback` with :meth:`~.Callback.on_criterion`
     calling the decorated function
 
-    :param func: The function(state) to *decorate*
-    :type func: function
-    :return: Initialised callback with :meth:`.Callback.on_criterion` calling func
-    :rtype: :class:`.Callback`
+    Args:
+        func (function): The function(state) to *decorate*
+
+    Returns:
+        :class:`.Callback`: Initialised callback with :meth:`~.Callback.on_criterion` calling func
     """
     return bind_to(Callback.on_criterion)(func)
 
@@ -100,10 +105,11 @@ def on_backward(func):
     """ The :func:`on_backward` decorator is used to initialise a :class:`.Callback` with :meth:`~.Callback.on_backward`
     calling the decorated function
 
-    :param func: The function(state) to *decorate*
-    :type func: function
-    :return: Initialised callback with :meth:`.Callback.on_backward` calling func
-    :rtype: :class:`.Callback`
+    Args:
+        func (function): The function(state) to *decorate*
+
+    Returns:
+        :class:`.Callback`: Initialised callback with :meth:`~.Callback.on_backward` calling func
     """
     return bind_to(Callback.on_backward)(func)
 
@@ -112,10 +118,11 @@ def on_step_training(func):
     """ The :func:`on_step_training` decorator is used to initialise a :class:`.Callback` with :meth:`~.Callback.on_step_training`
     calling the decorated function
 
-    :param func: The function(state) to *decorate*
-    :type func: function
-    :return: Initialised callback with :meth:`.Callback.on_step_training` calling func
-    :rtype: :class:`.Callback`
+    Args:
+        func (function): The function(state) to *decorate*
+
+    Returns:
+        :class:`.Callback`: Initialised callback with :meth:`~.Callback.on_step_training` calling func
     """
     return bind_to(Callback.on_step_training)(func)
 
@@ -124,10 +131,11 @@ def on_end_training(func):
     """ The :func:`on_end_training` decorator is used to initialise a :class:`.Callback` with :meth:`~.Callback.on_end_training`
     calling the decorated function
 
-    :param func: The function(state) to *decorate*
-    :type func: function
-    :return: Initialised callback with :meth:`.Callback.on_end_training` calling func
-    :rtype: :class:`.Callback`
+    Args:
+        func (function): The function(state) to *decorate*
+
+    Returns:
+        :class:`.Callback`: Initialised callback with :meth:`~.Callback.on_end_training` calling func
     """
     return bind_to(Callback.on_end_training)(func)
 
@@ -136,10 +144,11 @@ def on_end_epoch(func):
     """ The :func:`on_end_epoch` decorator is used to initialise a :class:`.Callback` with :meth:`~.Callback.on_end_epoch`
     calling the decorated function
 
-    :param func: The function(state) to *decorate*
-    :type func: function
-    :return: Initialised callback with :meth:`.Callback.on_end_epoch` calling func
-    :rtype: :class:`.Callback`
+    Args:
+        func (function): The function(state) to *decorate*
+
+    Returns:
+        :class:`.Callback`: Initialised callback with :meth:`~.Callback.on_end_epoch` calling func
     """
     return bind_to(Callback.on_end_epoch)(func)
 
@@ -148,10 +157,11 @@ def on_end(func):
     """ The :func:`on_end` decorator is used to initialise a :class:`.Callback` with :meth:`~.Callback.on_end`
     calling the decorated function
 
-    :param func: The function(state) to *decorate*
-    :type func: function
-    :return: Initialised callback with :meth:`.Callback.on_end` calling func
-    :rtype: :class:`.Callback`
+    Args:
+        func (function): The function(state) to *decorate*
+
+    Returns:
+        :class:`.Callback`: Initialised callback with :meth:`~.Callback.on_end` calling func
     """
     return bind_to(Callback.on_end)(func)
 
@@ -160,10 +170,11 @@ def on_start_validation(func):
     """ The :func:`on_start_validation` decorator is used to initialise a :class:`.Callback` with :meth:`~.Callback.on_start_validation`
     calling the decorated function
 
-    :param func: The function(state) to *decorate*
-    :type func: function
-    :return: Initialised callback with :meth:`.Callback.on_start_validation` calling func
-    :rtype: :class:`.Callback`
+    Args:
+        func (function): The function(state) to *decorate*
+
+    Returns:
+        :class:`.Callback`: Initialised callback with :meth:`~.Callback.on_start_validation` calling func
     """
     return bind_to(Callback.on_start_validation)(func)
 
@@ -172,10 +183,11 @@ def on_sample_validation(func):
     """ The :func:`on_sample_validation` decorator is used to initialise a :class:`.Callback` with :meth:`~.Callback.on_sample_validation`
     calling the decorated function
 
-    :param func: The function(state) to *decorate*
-    :type func: function
-    :return: Initialised callback with :meth:`.Callback.on_sample_validation` calling func
-    :rtype: :class:`.Callback`
+    Args:
+        func (function): The function(state) to *decorate*
+
+    Returns:
+        :class:`.Callback`: Initialised callback with :meth:`~.Callback.on_sample_validation` calling func
     """
     return bind_to(Callback.on_sample_validation)(func)
 
@@ -184,10 +196,11 @@ def on_forward_validation(func):
     """ The :func:`on_forward_validation` decorator is used to initialise a :class:`.Callback` with :meth:`~.Callback.on_forward_validation`
     calling the decorated function
 
-    :param func: The function(state) to *decorate*
-    :type func: function
-    :return: Initialised callback with :meth:`.Callback.on_forward_validation` calling func
-    :rtype: :class:`.Callback`
+    Args:
+        func (function): The function(state) to *decorate*
+
+    Returns:
+        :class:`.Callback`: Initialised callback with :meth:`~.Callback.on_forward_validation` calling func
     """
     return bind_to(Callback.on_forward_validation)(func)
 
@@ -196,10 +209,11 @@ def on_criterion_validation(func):
     """ The :func:`on_criterion_validation` decorator is used to initialise a :class:`.Callback` with :meth:`~.Callback.on_criterion_validation`
     calling the decorated function
 
-    :param func: The function(state) to *decorate*
-    :type func: function
-    :return: Initialised callback with :meth:`.Callback.on_criterion_validation` calling func
-    :rtype: :class:`.Callback`
+    Args:
+        func (function): The function(state) to *decorate*
+
+    Returns:
+        :class:`.Callback`: Initialised callback with :meth:`~.Callback.on_criterion_validation` calling func
     """
     return bind_to(Callback.on_criterion_validation)(func)
 
@@ -208,10 +222,11 @@ def on_end_validation(func):
     """ The :func:`on_end_validation` decorator is used to initialise a :class:`.Callback` with :meth:`~.Callback.on_end_validation`
     calling the decorated function
 
-    :param func: The function(state) to *decorate*
-    :type func: function
-    :return: Initialised callback with :meth:`.Callback.on_end_validation` calling func
-    :rtype: :class:`.Callback`
+    Args:
+        func (function): The function(state) to *decorate*
+
+    Returns:
+        :class:`.Callback`: Initialised callback with :meth:`~.Callback.on_end_validation` calling func
     """
     return bind_to(Callback.on_end_validation)(func)
 
@@ -220,10 +235,11 @@ def on_step_validation(func):
     """ The :func:`on_step_validation` decorator is used to initialise a :class:`.Callback` with :meth:`~.Callback.on_step_validation`
     calling the decorated function
 
-    :param func: The function(state) to *decorate*
-    :type func: function
-    :return: Initialised callback with :meth:`.Callback.on_step_validation` calling func
-    :rtype: :class:`.Callback`
+    Args:
+        func (function): The function(state) to *decorate*
+
+    Returns:
+        :class:`.Callback`: Initialised callback with :meth:`~.Callback.on_step_validation` calling func
     """
     return bind_to(Callback.on_step_validation)(func)
 
@@ -232,10 +248,11 @@ def add_to_loss(func):
     """ The :func:`add_to_loss` decorator is used to initialise a :class:`.Callback` with the value returned from func
     being added to the loss
 
-    :param func: The function(state) to *decorate*
-    :type func: function
-    :return: Initialised callback which adds the returned value from func to the loss
-    :rtype: :class:`.Callback`
+    Args:
+        func (function): The function(state) to *decorate*
+
+    Returns:
+        :class:`.Callback`: Initialised callback which adds the returned value from func to the loss
     """
     @on_criterion
     @on_criterion_validation
@@ -248,8 +265,12 @@ def add_to_loss(func):
 def once(fcn):
     """
     Decorator to fire a callback once in the entire fitting procedure.
-    :param fcn: the `torchbearer callback` function to decorate.
-    :return: the decorator
+
+    Args:
+        fcn (function): the `torchbearer callback` function to decorate.
+
+    Returns:
+        the decorator
     """
     done = False
 
@@ -266,8 +287,12 @@ def once(fcn):
 def once_per_epoch(fcn):
     """
     Decorator to fire a callback once (on the first call) in any given epoch.
-    :param fcn: the `torchbearer callback` function to decorate.
-    :return: the decorator
+
+    Args:
+        fcn (function): the `torchbearer callback` function to decorate.
+
+    Returns:
+        the decorator
     """
     last_epoch = None
 
@@ -284,11 +309,15 @@ def once_per_epoch(fcn):
 def only_if(condition_expr):
     """
     Decorator to fire a callback only if the given conditional expression function returns True.
-    :param condition_expr: a function/lambda that must evaluate to true for the\
-                           decorated `torchbearer callback` to be called. The `state`\
-                           object passed to the callback will be passed as an argument\
-                           to the condition function.
-    :return: the decorator
+
+    Args:
+        condition_expr: a function/lambda that must evaluate to true for the\
+                        decorated `torchbearer callback` to be called. The `state`\
+                        object passed to the callback will be passed as an argument\
+                        to the condition function.
+
+    Returns:
+        the decorator
     """
     def condition_decorator(fcn):
         if isinstance(fcn, LambdaCallback):