+++ torchbearer/callbacks/early_stopping.py
@@ -6,16 +6,18 @@ from torchbearer.callbacks import Callback
 class EarlyStopping(Callback):
     """Callback to stop training when a monitored quantity has stopped improving.
 
-    :param monitor: Quantity to be monitored
-    :type monitor: str
-    :param min_delta: Minimum change in the monitored quantity to qualify as an improvement, i.e. an absolute change of less than min_delta, will count as no improvement.
-    :type min_delta: float
-    :param patience: Number of epochs with no improvement after which training will be stopped.
-    :type patience: int
-    :param verbose: Verbosity mode, will print stopping info if verbose > 0
-    :type verbose: int
-    :param mode: One of {auto, min, max}. In `min` mode, training will stop when the quantity monitored has stopped decreasing; in `max` mode it will stop when the quantity monitored has stopped increasing; in `auto` mode, the direction is automatically inferred from the name of the monitored quantity.
-    :type mode: str
+    Args:
+        monitor (str): Name of quantity in metrics to be monitored
+        min_delta (float): Minimum change in the monitored quantity to qualify as an improvement, i.e. an absolute
+            change of less than min_delta, will count as no improvement.
+        patience (int): Number of epochs with no improvement after which training will be stopped.
+        verbose (int): Verbosity mode, will print stopping info if verbose > 0
+        mode (str): One of {auto, min, max}. In `min` mode, training will stop when the quantity monitored has stopped
+            decreasing; in `max` mode it will stop when the quantity monitored has stopped increasing; in `auto` mode,
+            the direction is automatically inferred from the name of the monitored quantity.
+
+    State Requirements:
+        - :attr:`torchbearer.state.METRICS`: Metrics should be a dict containing the given monitor key as a minimum
     """
 
     def __init__(self, monitor='val_loss',