+++ torchbearer/callbacks/terminate_on_nan.py
@@ -6,11 +6,13 @@ import math
 
 
 class TerminateOnNaN(Callback):
-    """Callback which montiors the given metric and halts training if its value is nan or
-    inf.
+    """Callback which montiors the given metric and halts training if its value is nan or inf.
 
-    :param monitor: The metric name to monitor
-    :type monitor: str
+    Args:
+        monitor (str): The name of the metric to monitor
+
+    State Requirements:
+        - :attr:`torchbearer.state.METRICS`: Metrics should be a dict containing at least the key `monitor`
     """
     def __init__(self, monitor='running_loss'):
         super(TerminateOnNaN, self).__init__()